ktra số

+++ src/Controller/CheckInController.java
@@ -34,7 +34,7 @@ public class CheckInController {
     }
     private boolean isNumberOrNot(String input) {
         try {
-            Integer.parseInt(input);
+            Long.parseLong(input);
         } catch (NumberFormatException e) {
             return false;
         }
