+++ src/Controller/DichVuController.java
@@ -26,7 +26,7 @@ public class DichVuController {
 
     private boolean isNumberOrNot(String input) {
         try {
-            Integer.parseInt(input);
+            Long.parseLong(input);
         } catch (NumberFormatException e) {
             return false;
         }
