+++ src/Controller/KieuPhongController.java
@@ -31,7 +31,7 @@ public class KieuPhongController {
     
     private boolean isNumberOrNot(String input) {
         try {
-            Integer.parseInt(input);
+            Long.parseLong(input);
         } catch (NumberFormatException e) {
             return false;
         }
